+++ front-end/src/pages/race/RaceHostingPage.vue
@@ -33,10 +33,10 @@
 import { defineComponent, onMounted, onUnmounted, ref } from 'vue';
 import { useRoute } from 'vue-router';
 import TextModal from '../../components/global/modals/TextModal.vue';
+import { useTextModal } from '../../components/global/modals/UseTextModal';
 import RaceUser from '../../components/race/RaceUser.vue';
 import router from '../../router/Router';
 import { raceSessionService } from '../../services/game/RaceSessionService';
-import { textModal } from '../../types/global/modals/TextModal';
 import { raceUser } from '../../types/race/RaceUser';
 import { useRaceAccessGuard } from './RaceAccessGuard';
 
@@ -45,19 +45,8 @@ export default defineComponent({
     components: { TextModal, RaceUser },
     setup() {
         useRaceAccessGuard();
-        let isTextModalVisible = ref(false)
-
-        let textModalContent = ref<textModal>({
-            header: 'Alert',
-            body: 'Nothing to show'
-        });
-
-        const showModal = (header: string, body: string) => {
-            textModalContent.value.header = header;
-            textModalContent.value.body = body;
-            isTextModalVisible.value = true;
-        };
-
+        const { isTextModalVisible, textModalContent, showModal } = useTextModal();
+        
         const route = useRoute();
         const raceId = Number(route.params.raceId);
         const sessionId = Number(route.params.sessionId);
